--- src/Layouts/Landing/index.ts
@@ -1,3 +1,0 @@
-import Landing from './Landing';
-
-export default Landing;
