--- src/Layouts/Landing/subcomponents/ExplaneCard/index.ts
@@ -1,3 +1,0 @@
-import ExplaneCard from './ExplaneCard';
-
-export default ExplaneCard;
