--- src/Layouts/Landing/subcomponents/PricingCard/index.ts
@@ -1,3 +1,0 @@
-import PricingCard from './PricingCard';
-
-export default PricingCard;
