+++ src/pages/auth/signIn/constants/index.ts
@@ -1,7 +1,18 @@
-import { validateEmail, validatePassword } from '../../utils/validation';
+import { validateEmail, validatePassword } from '@/utils/Auth/validation';
 
-export const SIGN_IN_INPUT_DATA = [
+interface SignInData {
+    action: 'signIn';
+    key: string;
+    label: string;
+    type: string;
+    placeholder: string;
+    validate: (value: string) => boolean;
+    inValidMessage: string;
+}
+
+export const SIGN_IN_INPUT_DATA: SignInData[] = [
     {
+        action: 'signIn',
         key: 'email',
         label: '이메일',
         type: 'text',
@@ -10,6 +21,7 @@ export const SIGN_IN_INPUT_DATA = [
         inValidMessage: '이메일 형식이 올바르지 않습니다.'
     },
     {
+        action: 'signIn',
         key: 'password',
         label: '패스워드',
         type: 'password',
