--- src/pages/auth/signIn/subcomponents/OAuthButton/index.ts
@@ -1,3 +1,0 @@
-import OAuthButton from './OAuthButton';
-
-export default OAuthButton;
