+++ src/pages/auth/signUp/constants/index.ts
@@ -2,10 +2,21 @@ import {
     validateNickname,
     validateEmail,
     validatePassword
-} from '../../utils/validation';
+} from '@/utils/Auth/validation';
 
-export const SIGN_UP_INPUT_DATA = [
+interface SignUpData {
+    action: 'signUp';
+    key: string;
+    label: string;
+    type: string;
+    placeholder: string;
+    validate: (value: string) => boolean;
+    inValidMessage: string;
+}
+
+export const SIGN_UP_INPUT_DATA: SignUpData[] = [
     {
+        action: 'signUp',
         key: 'nickname',
         label: '닉네임',
         type: 'text',
@@ -14,6 +25,7 @@ export const SIGN_UP_INPUT_DATA = [
         inValidMessage: '2~16자의 영문, 숫자, 한글만 사용 가능합니다.'
     },
     {
+        action: 'signUp',
         key: 'email',
         label: '이메일',
         type: 'text',
@@ -22,6 +34,7 @@ export const SIGN_UP_INPUT_DATA = [
         inValidMessage: '이메일 형식이 올바르지 않습니다.'
     },
     {
+        action: 'signUp',
         key: 'password',
         label: '비밀번호',
         type: 'password',
