--- src/pages/auth/signUp/subcomponents/SamePasswordCheckInput/index.ts
@@ -1,3 +1,0 @@
-import SamePasswordCheckInput from './SamePasswordCheckInput';
-
-export default SamePasswordCheckInput;
