--- src/pages/auth/subcomponents/AuthContainer/AuthContainer.style.tsx
@@ -1,7 +1,0 @@
-import { css } from '@emotion/react';
-
-export const authContainerStyle = css`
-    display: flex;
-    flex-direction: column;
-    flex: 1;
-`;
