--- src/pages/auth/subcomponents/AuthContainer/subcomponents/AuthHeader/AuthHeader.style.tsx
@@ -1,7 +1,0 @@
-import { css } from '@emotion/react';
-
-export const authHeaderStyle = css`
-    padding: 1rem 2rem;
-
-    cursor: pointer;
-`;
