--- src/pages/auth/subcomponents/AuthContainer/subcomponents/AuthHeader/index.ts
@@ -1,3 +1,0 @@
-import AuthHeader from './AuthHeader';
-
-export default AuthHeader;
