--- src/pages/auth/subcomponents/InputContainer/InputContainer.tsx
@@ -1,77 +1,0 @@
-import React from 'react';
-
-import { FlexContainer, Label, Input, Typography } from 'nimble-ds';
-
-import { commonSetInputTextWithKey } from '../../utils/common';
-
-import { IUserLogin, IUserSignup } from '@/types/user';
-
-type Props = {
-    id: string;
-    type: any;
-    placeholder: string;
-    labelText: string;
-    inValidMessage: string;
-    currentData: IUserLogin | IUserSignup;
-    handleChangeFunctions: Function;
-    validateFunction: Function;
-};
-
-const InputContainer = ({
-    id,
-    type,
-    placeholder,
-    labelText,
-    inValidMessage,
-    currentData,
-    handleChangeFunctions,
-    validateFunction
-}: Props) => {
-    const [isValid, setIsValid] = React.useState(true);
-    const [inputValue, setInputValue] = React.useState('');
-
-    const updateInputValue = (e: React.ChangeEvent<HTMLInputElement>) => {
-        setInputValue(e.target.value);
-
-        commonSetInputTextWithKey({
-            handleChangeFunctions,
-            currentData,
-            key: id,
-            value: e.target.value
-        });
-    };
-
-    const validateInputValue = (value: string) => {
-        const isValid = validateFunction(value);
-
-        setIsValid(isValid);
-    };
-
-    React.useEffect(() => {
-        if (!isValid) {
-            validateInputValue(inputValue);
-        }
-    }, [inputValue]);
-
-    return (
-        <FlexContainer direction="column" gap="0.5rem">
-            <Label htmlFor={id}>
-                <Typography value={labelText} size="14px" />
-            </Label>
-            <Input
-                id={id}
-                type={type}
-                value={inputValue}
-                onChange={updateInputValue}
-                onBlur={() => validateInputValue(inputValue)}
-                placeholder={placeholder}
-                size="lg"
-                width={256}
-                invalid={!isValid && inputValue.length !== 0}
-                invalidMessage={inValidMessage}
-            />
-        </FlexContainer>
-    );
-};
-
-export default InputContainer;
