--- src/pages/auth/subcomponents/InputContainer/index.ts
@@ -1,3 +1,0 @@
-import InputContainer from './InputContainer';
-
-export default InputContainer;
