--- src/pages/auth/subcomponents/ServiceInfoContainer/ServiceInfoContainer.style.tsx
@@ -1,8 +1,0 @@
-import { css } from '@emotion/react';
-
-export const serviceInfoContainerMainStyle = css`
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex: 1;
-`;
