--- src/pages/auth/subcomponents/ServiceInfoContainer/ServiceInfoContainer.tsx
@@ -1,11 +1,0 @@
-import { serviceInfoContainerMainStyle } from './ServiceInfoContainer.style';
-
-const ServiceInfoContainer = () => {
-    return (
-        <section css={serviceInfoContainerMainStyle}>
-            ServiceInfoContainer
-        </section>
-    );
-};
-
-export default ServiceInfoContainer;
