--- src/pages/auth/subcomponents/ServiceInfoContainer/index.ts
@@ -1,3 +1,0 @@
-import ServiceInfoContainer from './ServiceInfoContainer';
-
-export default ServiceInfoContainer;
