--- src/pages/main/components/CurrentTimeContainer/CurrentTimeContainer.style.tsx
@@ -1,25 +1,0 @@
-import { css } from '@emotion/react';
-
-export const currentTimeContainerStyle = css`
-    width: 25rem;
-    height: 30rem;
-
-    border-radius: 2rem;
-    border: 1px solid #d5d5d5;
-`;
-
-export const currentTimeTopStyle = css`
-    height: 10rem;
-
-    background-color: #d5d5d5;
-
-    border-top-left-radius: 2rem;
-    border-top-right-radius: 2rem;
-`;
-
-export const currentTimeBottomStyle = css`
-    height: 20rem;
-
-    border-bottom-left-radius: 2rem;
-    border-bottom-right-radius: 2rem;
-`;
