--- src/pages/main/components/CurrentTimeContainer/index.ts
@@ -1,3 +1,0 @@
-import CurrentTimeContainer from './CurrentTimeContainer';
-
-export default CurrentTimeContainer;
