--- src/pages/main/components/MainContainer/MainContainer.style.tsx
@@ -1,7 +1,0 @@
-import { css } from '@emotion/react';
-
-export const mainStyle = css`
-    width: 25rem;
-
-    padding: 1.5rem;
-`;
