--- src/pages/main/components/MainContainer/constants/index.ts
@@ -1,14 +1,0 @@
-export const ACTION_ITEMS = [
-    {
-        title: 'New Meeting'
-    },
-    {
-        title: 'Join'
-    },
-    {
-        title: 'Schedule'
-    },
-    {
-        title: 'Share Screen'
-    }
-];
