--- src/pages/main/components/MainContainer/index.ts
@@ -1,3 +1,0 @@
-import MainContainer from './MainContainer';
-
-export default MainContainer;
