--- src/pages/main/components/MainContainer/subcomponents/ActionButton/ActionButton.style.tsx
@@ -1,11 +1,0 @@
-import { css } from '@emotion/react';
-
-export const actionButtonStyle = css`
-    background-color: #ededed;
-    width: 10rem;
-    height: 10rem;
-
-    padding: 1rem;
-
-    border-radius: 2rem;
-`;
