--- src/pages/main/components/Navigation/Navigation.style.tsx
@@ -1,9 +1,0 @@
-import { css } from '@emotion/react';
-
-export const navStyle = css`
-    background-color: #ededed;
-    height: 6rem;
-    width: 100%;
-
-    padding: 1rem 12rem;
-`;
