--- src/pages/main/components/Navigation/constants/index.ts
@@ -1,14 +1,0 @@
-export const NAVIGATION_ITEMS = [
-    {
-        title: 'Home',
-        urn: '/main'
-    },
-    {
-        title: 'Chat',
-        urn: '/main/chat'
-    },
-    {
-        title: 'Meetings',
-        urn: '/main/meetings'
-    }
-];
