--- src/pages/main/components/Navigation/index.ts
@@ -1,3 +1,0 @@
-import Navigation from './Navigation';
-
-export default Navigation;
