--- src/pages/main/components/Navigation/subcomponents/NavigationButton/NavigationButton.style.tsx
@@ -1,6 +1,0 @@
-import { css } from '@emotion/react';
-
-export const navigationButtonStyle = css`
-    text-decoration: none;
-    color: #5a5a5a;
-`;
