--- src/pages/main/components/Navigation/subcomponents/NavigationButton/index.ts
@@ -1,3 +1,0 @@
-import NavigationButton from './NavigationButton';
-
-export default NavigationButton;
