--- src/types/user.ts
@@ -1,10 +1,0 @@
-export interface IUserLogin {
-    email: string;
-    password: string;
-}
-
-export interface IUserSignup {
-    nickname: string;
-    email: string;
-    password: string;
-}
