--- src/utils/typography.ts
@@ -1,114 +1,0 @@
-export const COLORS = {
-    GRAY_50: 'gray50',
-    GRAY_100: 'gray100',
-    GRAY_200: 'gray200',
-    GRAY_300: 'gray300',
-    GRAY_400: 'gray400',
-    GRAY_500: 'gray500',
-    GRAY_600: 'gray600',
-    GRAY_700: 'gray700',
-    GRAY_800: 'gray800',
-    GRAY_900: 'gray900',
-    BLUE_50: 'blue50',
-    BLUE_100: 'blue100',
-    BLUE_200: 'blue200',
-    BLUE_300: 'blue300',
-    BLUE_400: 'blue400',
-    BLUE_500: 'blue500',
-    BLUE_600: 'blue600',
-    BLUE_700: 'blue700',
-    BLUE_800: 'blue800',
-    BLUE_900: 'blue900',
-    RED_50: 'red50',
-    RED_100: 'red100',
-    RED_200: 'red200',
-    RED_300: 'red300',
-    RED_400: 'red400',
-    RED_500: 'red500',
-    RED_600: 'red600',
-    RED_700: 'red700',
-    RED_800: 'red800',
-    RED_900: 'red900',
-    ORANGE_50: 'orange50',
-    ORANGE_100: 'orange100',
-    ORANGE_200: 'orange200',
-    ORANGE_300: 'orange300',
-    ORANGE_400: 'orange400',
-    ORANGE_500: 'orange500',
-    ORANGE_600: 'orange600',
-    ORANGE_700: 'orange700',
-    ORANGE_800: 'orange800',
-    ORANGE_900: 'orange900',
-    GREEN_50: 'green50',
-    GREEN_100: 'green100',
-    GREEN_200: 'green200',
-    GREEN_300: 'green300',
-    GREEN_400: 'green400',
-    GREEN_500: 'green500',
-    GREEN_600: 'green600',
-    GREEN_700: 'green700',
-    GREEN_800: 'green800',
-    GREEN_900: 'green900',
-    GRAPE_50: 'grape50',
-    GRAPE_100: 'grape100',
-    GRAPE_200: 'grape200',
-    GRAPE_300: 'grape300',
-    GRAPE_400: 'grape400',
-    GRAPE_500: 'grape500',
-    GRAPE_600: 'grape600',
-    GRAPE_700: 'grape700',
-    GRAPE_800: 'grape800',
-    GRAPE_900: 'grape900',
-    VIOLET_50: 'violet50',
-    VIOLET_100: 'violet100',
-    VIOLET_200: 'violet200',
-    VIOLET_300: 'violet300',
-    VIOLET_400: 'violet400',
-    VIOLET_500: 'violet500',
-    VIOLET_600: 'violet600',
-    VIOLET_700: 'violet700',
-    VIOLET_800: 'violet800',
-    VIOLET_900: 'violet900',
-    CYAN_50: 'cyan50',
-    CYAN_100: 'cyan100',
-    CYAN_200: 'cyan200',
-    CYAN_300: 'cyan300',
-    CYAN_400: 'cyan400',
-    CYAN_500: 'cyan500',
-    CYAN_600: 'cyan600',
-    CYAN_700: 'cyan700',
-    CYAN_800: 'cyan800',
-    CYAN_900: 'cyan900',
-    LIME_50: 'lime50',
-    LIME_100: 'lime100',
-    LIME_200: 'lime200',
-    LIME_300: 'lime300',
-    LIME_400: 'lime400',
-    LIME_500: 'lime500',
-    LIME_600: 'lime600',
-    LIME_700: 'lime700',
-    LIME_800: 'lime800',
-    LIME_900: 'lime900',
-    YELLOW_50: 'yellow50',
-    YELLOW_100: 'yellow100',
-    YELLOW_200: 'yellow200',
-    YELLOW_300: 'yellow300',
-    YELLOW_400: 'yellow400',
-    YELLOW_500: 'yellow500',
-    YELLOW_600: 'yellow600',
-    YELLOW_700: 'yellow700',
-    YELLOW_800: 'yellow800',
-    YELLOW_900: 'yellow900'
-} as const;
-
-type Color = typeof COLORS[keyof typeof COLORS];
-
-type Weight = 'sm' | 'md' | 'lg';
-
-export function isColor(value: any): value is Color {
-    return Object.values(COLORS).includes(value);
-}
-
-export function isWeight(weight: string): weight is Weight {
-    return ['sm', 'md', 'lg'].includes(weight);
-}
